+++ src/components/Table/Table.test.tsx
@@ -15,7 +15,7 @@ type Props = React.ComponentProps<typeof Table>;
 /**
  * Mock ButtonSelect
  */
-jest.mock('../ButtonSelect');
+jest.mock('../ui/ButtonSelect');
 
 describe('Table', () => {
   /**
