+++ src/components/Table/Table.tsx
@@ -20,10 +20,10 @@ import {
 import { useVirtualizer } from '@tanstack/react-virtual';
 import React, { CSSProperties, MutableRefObject, RefObject, useCallback, useEffect, useMemo, useState } from 'react';
 
+import { ButtonSelect } from '@/components';
 import { TEST_IDS } from '@/constants';
 import { ColumnPinDirection, Pagination as PaginationOptions } from '@/types';
 
-import { ButtonSelect } from '../ButtonSelect';
 import { TableHeaderCell, TableRow } from './components';
 import { useEditableData, useSortState, useSyncedColumnFilters } from './hooks';
 import { getStyles } from './Table.styles';
