+++ src/components/editors/EditableDataEditor/EditableDataEditor.styles.ts
@@ -1,0 +1,10 @@
+import { css } from '@emotion/css';
+import { GrafanaTheme2 } from '@grafana/data';
+
+export const getStyles = (theme: GrafanaTheme2) => {
+  return {
+    item: css`
+      margin-bottom: ${theme.spacing(1)};
+    `,
+  };
+};
