+++ src/components/editors/EditableDataEditor/components/TableUpdateEditor/TableUpdateEditor.styles.ts
