+++ src/components/editors/EditableDataEditor/components/TableUpdateEditor/index.ts
@@ -1,0 +1,1 @@
+export * from './TableUpdateEditor';
