+++ src/components/editors/TablesEditor/TablesEditor.tsx
@@ -1,6 +1,6 @@
 import { cx } from '@emotion/css';
 import { StandardEditorProps } from '@grafana/data';
-import { Button, Icon, InlineField, InlineFieldRow, Input, useTheme2 } from '@grafana/ui';
+import { Alert, Button, Icon, InlineField, InlineFieldRow, Input, useTheme2 } from '@grafana/ui';
 import { DragDropContext, Draggable, DraggingStyle, Droppable, DropResult, NotDraggingStyle } from '@hello-pangea/dnd';
 import { Collapse } from '@volkovlabs/components';
 import React, { useCallback, useMemo, useState } from 'react';
@@ -26,6 +26,11 @@ const getItemStyle = (isDragging: boolean, draggableStyle: DraggingStyle | NotDr
    */
   ...draggableStyle,
 });
+
+/**
+ * Test Ids
+ */
+const testIds = TEST_IDS.tablesEditor;
 
 /**
  * Tables Editor
@@ -166,127 +171,135 @@ export const TablesEditor: React.FC<Props> = ({ context: { options, data }, onCh
 
   return (
     <>
-      <DragDropContext onDragEnd={onDragEnd}>
-        <Droppable droppableId="tables-editor">
-          {(provided) => (
-            <div {...provided.droppableProps} ref={provided.innerRef}>
-              {items.map((item, index) => (
-                <Draggable key={item.name} draggableId={item.name} index={index}>
-                  {(provided, snapshot) => (
-                    <div
-                      ref={provided.innerRef}
-                      {...provided.draggableProps}
-                      style={getItemStyle(snapshot.isDragging, provided.draggableProps.style)}
-                      className={styles.item}
-                    >
-                      <Collapse
-                        key={item.name}
-                        title={
-                          editItem === item.name ? (
-                            <div
-                              className={cx(styles.itemHeader, styles.itemHeaderForm)}
-                              onClick={(event) => event.stopPropagation()}
-                            >
-                              <InlineField className={styles.fieldName} invalid={!isUpdatedNameValid}>
-                                <Input
-                                  autoFocus={true}
-                                  value={editName}
-                                  onChange={(event) => setEditName(event.currentTarget.value)}
-                                  onKeyDown={(e) => {
-                                    if (e.key === 'Enter' && isUpdatedNameValid) {
-                                      onSaveName();
-                                    }
+      {items.length > 0 ? (
+        <DragDropContext onDragEnd={onDragEnd}>
+          <Droppable droppableId="tables-editor">
+            {(provided) => (
+              <div {...provided.droppableProps} ref={provided.innerRef}>
+                {items.map((item, index) => (
+                  <Draggable key={item.name} draggableId={item.name} index={index}>
+                    {(provided, snapshot) => (
+                      <div
+                        ref={provided.innerRef}
+                        {...provided.draggableProps}
+                        style={getItemStyle(snapshot.isDragging, provided.draggableProps.style)}
+                        className={styles.item}
+                      >
+                        <Collapse
+                          key={item.name}
+                          title={
+                            editItem === item.name ? (
+                              <div
+                                className={cx(styles.itemHeader, styles.itemHeaderForm)}
+                                onClick={(event) => event.stopPropagation()}
+                              >
+                                <InlineField className={styles.fieldName} invalid={!isUpdatedNameValid}>
+                                  <Input
+                                    autoFocus={true}
+                                    value={editName}
+                                    onChange={(event) => setEditName(event.currentTarget.value)}
+                                    onKeyDown={(e) => {
+                                      if (e.key === 'Enter' && isUpdatedNameValid) {
+                                        onSaveName();
+                                      }
 
-                                    if (e.key === 'Escape') {
-                                      onCancelEdit();
-                                    }
-                                  }}
-                                  {...TEST_IDS.tablesEditor.fieldName.apply()}
+                                      if (e.key === 'Escape') {
+                                        onCancelEdit();
+                                      }
+                                    }}
+                                    {...testIds.fieldName.apply()}
+                                  />
+                                </InlineField>
+                                <Button
+                                  variant="secondary"
+                                  fill="text"
+                                  className={styles.actionButton}
+                                  icon="times"
+                                  size="sm"
+                                  onClick={onCancelEdit}
+                                  {...testIds.buttonCancelRename.apply()}
                                 />
-                              </InlineField>
+                                <Button
+                                  variant="secondary"
+                                  fill="text"
+                                  className={styles.actionButton}
+                                  icon="save"
+                                  size="sm"
+                                  onClick={onSaveName}
+                                  disabled={!isUpdatedNameValid}
+                                  tooltip={
+                                    isUpdatedNameValid
+                                      ? ''
+                                      : 'Name is empty or table with the same name already exists.'
+                                  }
+                                  {...testIds.buttonSaveRename.apply()}
+                                />
+                              </div>
+                            ) : (
+                              <div className={cx(styles.itemHeader, styles.itemHeaderText)}>{item.name}</div>
+                            )
+                          }
+                          headerTestId={testIds.itemHeader.selector(item.name)}
+                          contentTestId={testIds.itemContent.selector(item.name)}
+                          actions={
+                            <>
+                              {editItem !== item.name && (
+                                <Button
+                                  icon="edit"
+                                  variant="secondary"
+                                  fill="text"
+                                  size="sm"
+                                  className={styles.actionButton}
+                                  onClick={() => {
+                                    /**
+                                     * Start Edit
+                                     */
+                                    setEditName(item.name);
+                                    setEditItem(item.name);
+                                  }}
+                                  {...testIds.buttonStartRename.apply()}
+                                />
+                              )}
                               <Button
-                                variant="secondary"
-                                fill="text"
-                                className={styles.actionButton}
-                                icon="times"
-                                size="sm"
-                                onClick={onCancelEdit}
-                                {...TEST_IDS.tablesEditor.buttonCancelRename.apply()}
-                              />
-                              <Button
-                                variant="secondary"
-                                fill="text"
-                                className={styles.actionButton}
-                                icon="save"
-                                size="sm"
-                                onClick={onSaveName}
-                                disabled={!isUpdatedNameValid}
-                                tooltip={
-                                  isUpdatedNameValid ? '' : 'Name is empty or table with the same name already exists.'
-                                }
-                                {...TEST_IDS.tablesEditor.buttonSaveRename.apply()}
-                              />
-                            </div>
-                          ) : (
-                            <div className={cx(styles.itemHeader, styles.itemHeaderText)}>{item.name}</div>
-                          )
-                        }
-                        headerTestId={TEST_IDS.tablesEditor.itemHeader.selector(item.name)}
-                        contentTestId={TEST_IDS.tablesEditor.itemContent.selector(item.name)}
-                        actions={
-                          <>
-                            {editItem !== item.name && (
-                              <Button
-                                icon="edit"
+                                icon="trash-alt"
                                 variant="secondary"
                                 fill="text"
                                 size="sm"
                                 className={styles.actionButton}
                                 onClick={() => {
                                   /**
-                                   * Start Edit
+                                   * Remove Item
                                    */
-                                  setEditName(item.name);
-                                  setEditItem(item.name);
+                                  onChangeItems(items.filter((column) => column.name !== item.name));
                                 }}
-                                {...TEST_IDS.tablesEditor.buttonStartRename.apply()}
+                                {...testIds.buttonRemove.apply()}
                               />
-                            )}
-                            <Button
-                              icon="trash-alt"
-                              variant="secondary"
-                              fill="text"
-                              size="sm"
-                              className={styles.actionButton}
-                              onClick={() => {
-                                /**
-                                 * Remove Item
-                                 */
-                                onChangeItems(items.filter((column) => column.name !== item.name));
-                              }}
-                              {...TEST_IDS.tablesEditor.buttonRemove.apply()}
-                            />
-                            <div className={styles.dragHandle} {...provided.dragHandleProps}>
-                              <Icon name="draggabledots" className={styles.dragIcon} />
-                            </div>
-                          </>
-                        }
-                        isOpen={collapseState[item.name]}
-                        onToggle={() => onToggleItemExpandedState(item.name)}
-                      >
-                        <TableEditor value={item} onChange={onChangeItem} data={data} />
-                      </Collapse>
-                    </div>
-                  )}
-                </Draggable>
-              ))}
-              {provided.placeholder}
-            </div>
-          )}
-        </Droppable>
-      </DragDropContext>
+                              <div className={styles.dragHandle} {...provided.dragHandleProps}>
+                                <Icon name="draggabledots" className={styles.dragIcon} />
+                              </div>
+                            </>
+                          }
+                          isOpen={collapseState[item.name]}
+                          onToggle={() => onToggleItemExpandedState(item.name)}
+                        >
+                          <TableEditor value={item} onChange={onChangeItem} data={data} />
+                        </Collapse>
+                      </div>
+                    )}
+                  </Draggable>
+                ))}
+                {provided.placeholder}
+              </div>
+            )}
+          </Droppable>
+        </DragDropContext>
+      ) : (
+        <Alert severity="info" title="No Tables" {...testIds.noItemsMessage.apply()}>
+          Please add at least one table to proceed.
+        </Alert>
+      )}
 
-      <InlineFieldRow className={styles.newItem} {...TEST_IDS.tablesEditor.newItem.apply()}>
+      <InlineFieldRow className={styles.newItem} {...testIds.newItem.apply()}>
         <InlineField
           label="New Table"
           grow={true}
@@ -297,7 +310,7 @@ export const TablesEditor: React.FC<Props> = ({ context: { options, data }, onCh
             placeholder="Unique name"
             value={newItemName}
             onChange={(event) => setNewItemName(event.currentTarget.value.trim())}
-            {...TEST_IDS.tablesEditor.newItemName.apply()}
+            {...testIds.newItemName.apply()}
           />
         </InlineField>
         <Button
@@ -305,7 +318,7 @@ export const TablesEditor: React.FC<Props> = ({ context: { options, data }, onCh
           title="Add Table"
           disabled={!newItemName || isNameExistsError}
           onClick={onAddNewItem}
-          {...TEST_IDS.tablesEditor.buttonAddNew.apply()}
+          {...testIds.buttonAddNew.apply()}
         >
           Add
         </Button>
