+++ src/components/editors/TablesEditor/components/TableEditor/components/index.ts
@@ -1,2 +1,1 @@
 export * from './ColumnsEditor';
-export * from './TableUpdateEditor';
