+++ src/components/editors/index.ts
@@ -1,5 +1,6 @@
 export * from './DatasourceEditor';
 export * from './DatasourcePayloadEditor';
 export * from './EditableColumnEditor';
+export * from './EditableDataEditor';
 export * from './FieldPicker';
 export * from './TablesEditor';
