+++ src/components/index.ts
@@ -1,4 +1,3 @@
-export * from './ButtonSelect';
 export * from './editors';
 export * from './Table';
 export * from './TablePanel';
