+++ src/components/ui/index.ts
@@ -1,2 +1,3 @@
+export * from './ButtonSelect';
 export * from './CollapseTitle';
 export * from './FieldsGroup';
